Move mapping/mapping_visu.py back to oak/mapping_visu.py
