+++ python/mapping/record_and_process_oak_d.sh
@@ -7,7 +7,7 @@ set -eux
 mkdir -p "$OUTPUT_FOLDER"
 
 echo "------------- Press Q to stop recording -------------"
-python mapping_visu.py --keyFrameCandidateInterval=4 --recordingFolder="$OUTPUT_FOLDER"
+python ../oak/mapping_visu.py --keyFrameCandidateInterval=4 --recordingFolder="$OUTPUT_FOLDER"
 python replay_to_nerf.py "$OUTPUT_FOLDER" "$OUTPUT_FOLDER"/nerfstudio --device_preset=oak-d --preview
 
 echo "Now run this Nerfstudio command:
